chore(ci): make clippy blocking; fix lints, await-holding-lock, advisories
Re-tightens the clippy gate (was informational since the CI-fix PR) and cleans
up the warnings behind it.

Clippy:
- cargo clippy --fix across the workspace cleared ~140 machine-applicable lints
  (redundant closures, needless borrows, collapsible ifs, unused imports, etc.).
- Fixed 6 `await_holding_lock` sites in the pool (aggregator/distributor start
  + aggregator/health/concurrency stop took/collected handles before awaiting;
  metrics server documented as intentional). These were real "guard held across
  await" hazards in the shutdown path.
- Fixed sort_by (-> sort_by_key + Reverse), field-reassign-after-default, an
  assertions-on-constants test, and removed dead locals/imports.
- Fixed a stale omnitak-cot test that constructed `Detail` without all fields
  (used ..Default::default()); it never compiled under --all-targets.
- CI clippy is now blocking: `-D warnings` with `-A dead_code` (intentional
  scaffolding) and a few complexity lints allowed as tracked tech-debt. The
  desktop-only GUI crate is excluded (no headless winit backend).

Security:
- `cargo update` (semver-compatible) cut RustSec advisories 128 -> 48, fixing
  the directly-fixable high-severity ones: aws-lc-sys TLS/PKCS7/X.509 bypasses
  (RUSTSEC-2026-0044..0048), bytes overflow, quinn, rustls-webpki, etc.
- The residual 48 are wasmtime (WASM plugin runtime; needs a major upgrade +
  API migration) and an ancient transitive tokio-0.1 chain. cargo audit stays
  informational (non-blocking) since blocking on unfixable transitive CVEs
  would keep CI permanently red.

Verified: fmt clean, headless clippy clean under policy, headless + GUI builds
compile, and the test suite passes.

+++ crates/omnitak-filter/benches/filter_bench.rs
@@ -173,7 +173,7 @@ fn bench_routing(c: &mut Criterion) {
             format!("Route {}", i),
             Arc::new(AffiliationFilter::friendly_only()),
             vec![format!("dest_{}", i)],
-            100 - i as i32,
+            100 - i,
         ));
     }
